+++ pytorch_lightning/trainer/configuration_validator.py
@@ -43,6 +43,8 @@ class ConfigValidator:
         elif self.trainer.state.fn == TrainerFn.PREDICTING:
             self.__verify_predict_loop_configuration(model)
         self.__verify_dp_batch_transfer_support(model)
+        # TODO(@daniellepintz): Delete _check_progress_bar in v1.7
+        self._check_progress_bar(model)
         # TODO: Delete _check_on_keyboard_interrupt in v1.7
         self._check_on_keyboard_interrupt()
 
@@ -109,6 +111,19 @@ class ConfigValidator:
                 "When using `Trainer(accumulate_grad_batches != 1)` and overriding"
                 "`LightningModule.optimizer_{step,zero_grad}`, the hooks will not be called on every batch"
                 "(rather, they are called on every optimization step)."
+            )
+
+    def _check_progress_bar(self, model: "pl.LightningModule") -> None:
+        r"""
+        Checks if get_progress_bar_dict is overriden and sends a deprecation warning.
+
+        Args:
+            model: The model to check the get_progress_bar_dict method.
+        """
+        if is_overridden("get_progress_bar_dict", model):
+            rank_zero_deprecation(
+                "The `LightningModule.get_progress_bar_dict` method was deprecated in v1.5 and will be removed in v1.7."
+                " Please use the `ProgressBarBase.get_metrics` instead."
             )
 
     def __verify_eval_loop_configuration(self, model: "pl.LightningModule", stage: str) -> None: